feat: add jpa/orm repo

+++ src/main/java/ru/vk/recommender/demospringitmo/workshop/controllers/dto/TaskStatsDto.java
@@ -1,8 +1,8 @@
 package ru.vk.recommender.demospringitmo.workshop.controllers.dto;
 
 public record TaskStatsDto(
-        int total,
-        int done,
-        int pending
+        long total,
+        long done,
+        long pending
 ) {
 }
